+++ src/main/kotlin/de/jvstvshd/chillingfoxes/foxesbot/config/Config.kt
@@ -20,7 +20,7 @@ import kotlin.io.path.outputStream
 class Config(private val path: Path = Path.of("config.json")) {
 
     companion object {
-        const val configVersion: Int = 0
+        const val configVersion: Int = 1
     }
 
     private val json: Json = Json {
@@ -41,11 +41,11 @@ class Config(private val path: Path = Path.of("config.json")) {
         }
         if (!Files.exists(path)) {
             Files.createFile(path)
-            configData = ConfigData(DataBaseData(),)
+            configData = ConfigData(DataBaseData())
             save()
         }
         if (path.fileSize() <= 0) {
-            configData = ConfigData(DataBaseData(),)
+            configData = ConfigData(DataBaseData())
             save()
         }
     }